refactor(sync): extract pairing URI codec into peers/codec.rs
